Separate the Conditions and Effects
