+++ Assets/Scripts/Bartering Minigame/Effect Cards/ConditionsAndEffects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dadfc48688d08774fa0b102813ec97ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
